requirements.txt
aiohttp==3.7.4.post0
aiosignal==1.2.0
astroid==2.5.1
async-timeout==3.0.1
asynctest==0.13.0
attrs==20.3.0
cachetools==4.2.1
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
charset-normalizer==2.0.12
click==7.1.2
discord.py==1.7.3
frozenlist==1.3.0
google-api-core==1.26.2
google-auth==1.28.0
google-cloud-texttospeech==2.2.0
googleapis-common-protos==1.53.0
grpcio==1.36.1
gTTS==2.2.2
idna==2.10
isort==5.8.0
lazy-object-proxy==1.6.0
libcst==0.3.17
mccabe==0.6.1
multidict==5.1.0
mypy-extensions==0.4.3
packaging==20.9
proto-plus==1.18.1
protobuf==3.15.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyNaCl==1.4.0
pyparsing==2.4.7
pytz==2021.1
PyYAML==5.4.1
requests==2.25.1
rsa==4.7.2
six==1.15.0
toml==0.10.2
typed-ast==1.4.2
typing-extensions==3.7.4.3
typing-inspect==0.6.0
urllib3==1.26.9
wrapt==1.12.1
yapf==0.31.0
yarl==1.6.3
